refactor: separate content input item
