+++ components/InputBar/index.js
@@ -1,3 +1,5 @@
+import createContent from './Items/content.js';
+
 function createInputBar() {
     const container = document.createElement('div');
     container.className = 'input-bar-container';
@@ -40,26 +42,7 @@ function createInputBar() {
     `;
     container.appendChild(amountItem);
 
-    const contentItem = document.createElement('div');
-    contentItem.className = 'input-bar-item';
-    contentItem.innerHTML = `
-        <div class="input-bar-item-wrapper">
-            <label for="content" class="light-12">내용</label>
-            <div class="content-length">
-                <span class="light-12 content-length-count">0</span>
-                <span class="light-12">/32</span>
-            </div>
-        </div>
-        <div class="input-bar-item-wrapper">
-            <input
-                type="text"
-                maxlength="32"
-                id="content"
-                class="semibold-12"
-                placeholder="입력하세요"
-            />
-        </div>
-    `;
+    const contentItem = createContent();
     container.appendChild(contentItem);
 
     const paymentMethodItem = document.createElement('div');
